+++ server/src/storage-config/storage-config.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { StorageConfigController } from './storage-config.controller';
+import { StorageConfigService } from './storage-config.service';
+
+@Module({
+  controllers: [StorageConfigController],
+  providers: [StorageConfigService],
+  exports: [StorageConfigService],
+})
+export class StorageConfigModule {}